+++ bin/gen-simple-ts-doc/markdown.d.ts
@@ -1,7 +1,0 @@
-export declare const Language: {
-    TS: string;
-    JS: string;
-};
-export declare function addCode(code: string, language?: string): string;
-export declare function addTitle(title: string, level?: number): string;
-export declare function getTitlePrefix(level: number): string;