+++ utils/index.d.ts
@@ -1,4 +1,0 @@
-export * from './src/utils';
-export * from './src/load';
-export * from './src/color';
-export { default as EventEmitter } from './src/event-emitter';