+++ utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zj-library/utils
 
+## 1.0.2
+
+### Patch Changes
+
+-   a9545a7: 删除.ts后缀
+
 ## 1.0.1
 
 ### Patch Changes