+++ vue-components/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @zj-library/vue-components
 
+## 1.0.3
+
+### Patch Changes
+
+-   a9545a7: 删除.ts后缀
+-   Updated dependencies [a9545a7]
+    -   @zj-library/vue-hooks@1.0.2
+    -   @zj-library/utils@1.0.2
+
 ## 1.0.2
 
 ### Patch Changes